requirements.txt
streamlit
pandas
python-dotenv
requests
beautifulsoup4
pydantic
tenacity
huggingface_hub
striprtf
pypdf2
openai